+++ docs/CHANGELOG.rst
@@ -6,6 +6,8 @@ Version 0.17.0 (unreleased)
 
 * Fix support for Vagrant >= 1.3.0
 * Ubuntu PPA fixes (Anthony Scalisi)
+* Update files.upload_template to match Fabric parameters
+  (thanks to Adam Patterson)
 
 
 Version 0.16.0 (2013-10-26)